# spec/api.yaml
openapi: 3.0.0
info:
   title: Korku Blog API Docs
   version: 1.0.0
   description: Endpoints documentation for the blog API 
paths:
...
   /auth/login:
      post:
         summary: 'Endpoint for user to log in'
         tags:
            - auth
         operationId: LoginRequest
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/LoginRequest'
            required: true
         responses:
            '200':
               description: 'success'
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/AuthResponse'
            '400':
               description: 'error'
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
components:
...
   schemas:
...
      LoginRequest:
         additionalProperties: false
         title: LoginRequest
         example: { email: 'kadepoju@mail.com', password: '1234abcd' }
         type: object
         properties:
            password:
               type: string
            email:
               type: string
            required:
               - password
               - email
...
      AuthResponse:
         title: AuthResponse
         type: object
         properties:
            success:
               type: string
            message:
               type: string
            status:
               type: number
            accessToken:
               type: string
         example:
            {
               success: 'true',
               message: 'operation successful',
               status: 200,
               accessToken: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImttYW5AZ21haWwuY29tIiwidHlwZSI6Ik5FVyBVU0VSIiwiaWF0IjoxNzE5NDYxNjM0LCJleHAiOjE3MTk0NjUyMzR9.DX9BlJpy2p8s5dOe1XESaZPqqu4fv7nCio8Qxex5YiM',
            }
      ErrorResponse:
         title: DefaultResponse
         type: object
         properties:
            error:
               type: string
            message:
               type: string
         required:
            - message
         example: { 'message': 'Error' }
...